Support self signed certificate chain (#989)
* fetch_tsa_certs: Refactor to enable self-signed chain

This change does not affect functionality yet (apart from slightly
changing the error messages if --gcp-ca-parent is not provided):
it's just refactoring to enable future work.

Most importantly naming of certificates and keys in the
fetchCertificateChain method is now consistent:
* "leaf" refers to the actual timestamp signing key/cert
* "parent" is the key/cert that signs "leaf": it may be a self-signed
  certicifate or an intermediate signed by CA
* "root" is a real CA certificate that signs "parent" (if a CA is used)

Signed-off-by: Jussi Kukkonen <jkukkonen@google.com>

* fetch_tsa_certs: Implement self-signed certificate chain

It is now possible to create a certificate chain that does not use a
real CA but instead has a self-signed signing certificate as the parent
of of the timestamp signing certificate.

This is a little experimental (e.g. cert lifetime is just hard coded now).

Signed-off-by: Jussi Kukkonen <jkukkonen@google.com>

* Add --intermediate-validity flag

This is useful mostly for the self signed case

Signed-off-by: Jussi Kukkonen <jkukkonen@google.com>

* fetch_tsa_certs: set additional certificate arguments

These are needed to get BasicConstraints in the cert.

Signed-off-by: Jussi Kukkonen <jkukkonen@google.com>

* README: Rewrite examples to include self-signed certificate

Signed-off-by: Jussi Kukkonen <jkukkonen@google.com>

* fetch-tsa-certs: Rename "intermediate" to parent

This changes one option that existed before this PR:
"--intermediate-kms-resource" is now "--parent-kms-resource".

The idea is that now this works in boths cases:
* parent is used as intermediate if a CA is provided
* parent is used as a self signed root if a CA is not provided

This also sorts the options the same way everywhere: leaf options, then
parent options, then CA root options, finally output options.

Signed-off-by: Jussi Kukkonen <jkukkonen@google.com>

* fetch-tsa-certs: Use validity argument for self-signed cert

Signed-off-by: Jussi Kukkonen <jkukkonen@google.com>

* fetch-tsa-certs: Uncapitalize local symbol

Signed-off-by: Jussi Kukkonen <jkukkonen@google.com>

---------

Signed-off-by: Jussi Kukkonen <jkukkonen@google.com>