Allow full issuing chain in response (#1082)
Fixes #1079

Per RFC3161, when the certReq field is set to true, the TSA's
certificate will be present in the timestamp response, and optionally
other certificates may be present. Other public TSAs provide the full
issuing chain in the response.

This PR adds a server configuration flag to include the full chain in
the response if the certReq bit is true.

Signed-off-by: Hayden B <8418760+haydentherapper@users.noreply.github.com>